// frontApp/src/app/componentes/models/solicitudAdopcion.ts
import { Animal } from './animal';


export class SolicitudAdopcion {
    id: number;
    cedula: string;
    nombres: string;
    apellidos: string;
    genero: string;
    edad: number;
    correo: string;
    telefono: string;
    celular: string;
    direccion: string;
    ciudad: string;
    mascota: string;
}